// MegabinFE/src/app/services/model/createScheduleContract.ts
/**
 * MySmartCloset API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DayOfWeek } from './dayOfWeek';
import { Frequency } from './frequency';

export interface CreateScheduleContract {
  addressId: string;
  dayOfWeek: DayOfWeek;
  frequency: Frequency;
}
export namespace CreateScheduleContract {}